Try fix

+++ flake.nix
@@ -1,13 +1,9 @@
 {
-  description = "Dustin's NixOS and MacOS configuration";
+  description = "Dustin's Configuration for NixOS and MacOS";
 
   inputs = {
-    nixpkgs = {
-      url = "github:dustinlyons/nixpkgs/master";
-    };
-    home-manager = {
-      url = "github:nix-community/home-manager";
-    };
+    nixpkgs.url = "github:NixOS/nixpkgs/master"; # Using the official nixpkgs repo
+    home-manager.url = "github:nix-community/home-manager";
     darwin = {
       url = "github:LnL7/nix-darwin/master";
       inputs.nixpkgs.follows = "nixpkgs";
@@ -18,25 +14,18 @@
     };
   };
 
-  outputs = { self, flake-utils, darwin, home-manager, nixpkgs, disko, ... }@inputs: {
+  outputs = { self, darwin, home-manager, nixpkgs, disko, ... }@inputs: {
 
     darwinConfigurations = {
       "Dustins-MBP" = darwin.lib.darwinSystem {
         system = "aarch64-darwin";
-        modules = [
-          ./darwin
-        ];
-        inputs = { inherit darwin home-manager nixpkgs; };
+        modules = [ ./darwin ];
       };
     };
 
-    nixosConfigurations = let
-      system = "x86_64-linux"; 
-      pkgs = nixpkgs.legacyPackages.${system};
-
-      felixDefault = {
-        inputs = { inherit nixpkgs; };
-        system = system;
+    nixosConfigurations = {
+      felix = {
+        system = "x86_64-linux";
         modules = [
           ./nixos
           disko.nixosModules.disko
@@ -47,27 +36,24 @@
           }
         ];
       };
-
-      nixosConfig = {
-        felix = nixpkgs.lib.nixosSystem {
-          inherit (felixDefault) system;
-          modules = felixDefault.modules ++ [];
-        };
-      };
-
-    in nixosConfig;
+    };
 
     apps = {
       x86_64-linux.bootstrap = {
         type = "app";
-        program = "${(pkgs.writeShellScriptBin "bootstrap-nixos" ''
-          sudo nix run ${disko} run-command -- --mode zap_create_mount --flake ${self}#nixosConfig.felix
+        program = "${(nixpkgs.legacyPackages.x86_64-linux.writeShellScriptBin "bootstrap-nixos" ''
+          set -e
+          sudo nix run ${disko} run-command -- --mode zap_create_mount --flake ${self}#nixosConfigurations.felix
 
           mkdir -p ~/.local/share/src/
-          git clone https://github.com/dustinlyons/nixos-config ~/.local/share/src/nixos-config
+          if ! git clone https://github.com/dustinlyons/nixos-config ~/.local/share/src/nixos-config; then
+            echo "Git clone failed!" >&2
+            exit 1
+          fi
+          
           ln -s ~/.local/share/src/nixos-config/flake.nix /mnt/etc/nixos/flake.nix
 
-          sudo nixos-install --flake /mnt/etc/nixos/#nixosConfig.felix
+          sudo nixos-install --flake /mnt/etc/nixos/#nixosConfigurations.felix
           reboot
         '')}/bin/bootstrap-nixos";
       };
